diariesとpostsで使うメソッドをutilにまとめ
